Updates for changes to other volatile-testing tools.
Pass --vars-file instead of --address-file to the volatile-address
generator.

Pass -DNOT_PRINT_CHECKSUM when compiling the test programs.  The Pin
tool now extracts and prints the checksum by itself.  This change
eliminates our dependency on printf.

And yesterday, we found a bug in printf (in glibc 2.15) that affected
our volatile checking!