Add new reduction helper scripts.
"prep_reduction.sh" prepares a file for reduction by preprocessing it.
We do this because the volatile-checking tool now rejects programs that
contain preprocessor directives.  It never reliably worked on
unpreprocessed sources; now it rejects such programs rather than
silently trying to carry on.

"showme.sh" is simply a copy of "test_reduction.sh", but with the noisy
and leave-your-trash flags turned on.  I've been using this to debug and
to see the results of the initial program and final reduced program.
But really, it needs to be merged into test_reduction somehow.