Implement argument pass-through to "volatile_test.pl".
All command-line arguments after "--" are passed through to the invocation of
the "volatile_test.pl" script.  For example,

  launchn.pl -- --no-vol-struct-union-fields --strict-volatile-rule

The pass-through is done sloppily, quoting-wise.  For example, if you were
crazy enough to do the following, you would get unexpected results:

  launchn.pl -- '*'

So don't do that.

This commit also removes the special "--use-pin-checksums" option, which was
just a pass-through anyway.  The new syntax for invoking this option is:

  launchn.pl -- --use-pin-checksums