Use new flags to support Pin-based checksumming.
This commit updates the command lines so that Pin-based checksumming
works again.  This catches up with command-line changes that Yang Chen
made some time ago.

But this commit is incomplete/wrong.  In particular, we need to compute
the value checksum over only those globals that appear in both of the
compiled programs.  The script does not yet do this.