Improve processing of var-addr files; still incomplete.
This commit (1) causes `sort' to pay attention to the first field only
and (2) causes `join' to output the fifth record field, which appears in
some of these address maps.

But it turns out that even with these changes, the code is still not
right.

It turns out that the first field is not a unique key when a program
uses packed structs.  One must deal with stuff like this:

     (g_172+0).f0; 0x611378; 3; non-pointer
     (g_172+0).f0; 0x61137b; 1; non-pointer; ff

which represents the locations within g_172[0], which is a structure
containing bitfields that span a 32-bit word.  Blech.

I'm just going to have to write the map-intersection by hand.