Update setup scripts for unified git repository.
The Emulab experiment-setup scripts now handle the new locations of
`volatile_checker' and `volatile_pintrace'.  These tools are now in this
git repository, not in separate SVN repositories.

This consolidation led to a sort of chicken-and-egg problem, which I
resolved by requiring experiments to start with a bootstrap script.
Read on for details.

One almost always wants to build programs from node-local sources, so
now we must clone this repository onto the Emulab node in order to build
volatile_checker and volatile_pintrace.  Where would we put the scripts
for this?  Well, in this repository, obviously :-).

(Previously we didn't bother to clone the Emulab-script repository; we
just assumed that a clone of this repository would be available to our
Emulab experiments via NFS.)

Now we create experiments that automatically clone this repository via
nodes' start command.  The start command runs a script, located in the
Emulab project's `/proj' tree, whose task is to clone this repository to
the nodes' local storage.  The script is expected to be a renamed copy
of `scripts/bootstrap.sh'.  See the NS files.

On the one hand, this seems clumsy and cumbersome.  On the other, the
introduction of a per-experiment bootstrap script makes it easier to do
certain things, such as setting the clone source and other details on a
per-experiment basis.  (The bootstrap scripts are named per experiment.)
More setup could be automated in this fashion, e.g., setting various
envvars or configuring the compilers under test.